pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dice-game"
version = "0.6.0"
description = "A simple CLI dice game"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
    { name = "Dennis Major",email = 'dennismajor0@gmail.com'}
]

maintainers = [
    { name = "Dennis Major",email = "dennismajor0@gmail.com"}
]

dependencies = [
    "argon2-cffi>=23.1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0"
]

keywords = ["game", "dice", "cli"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console"
]

[project.scripts]
dicegame = "dicegame.__main__:main"
